+++ react/localState/queries.js
@@ -3,90 +3,8 @@ import gql from 'graphql-tag'
 export const fullMinicartQuery = gql`
   query {
     minicart @client {
-      items {
-        upToDate
-        id
-        name
-        imageUrl
-        detailUrl
-        skuName
-        quantity
-        sellingPrice
-        listPrice
-        seller
-        index
-        parentItemIndex
-        parentAssemblyBinding
-        options {
-          seller
-          assemblyId
-          quantity
-          id
-        }
-        assemblyOptions {
-          parentPrice
-          added {
-            normalizedQuantity
-            choiceType
-            extraQuantity
-            item {
-              name
-              sellingPrice
-              quantity
-            }
-          }
-          removed {
-            removedQuantity
-            initialQuantity
-            name
-          }
-        }
-      }
-      orderForm {
-        cacheId
-        orderFormId
-        value
-        totalizers {
-          id
-          name
-          value
-        }
-        shippingData {
-          address {
-            id
-            neighborhood
-            complement
-            number
-            street
-            postalCode
-            city
-            reference
-            addressName
-            addressType
-          }
-          availableAddresses {
-            id
-            neighborhood
-            complement
-            number
-            street
-            postalCode
-            city
-            reference
-            addressName
-            addressType
-          }
-        }
-        clientProfileData {
-          email
-          firstName
-        }
-        storePreferencesData {
-          countryCode
-          currencyCode
-          timeZone
-        }
-      }
+      items
+      orderForm
     }
   }
 `
@@ -94,51 +12,7 @@ export const fullMinicartQuery = gql`
 export const minicartOrderFormQuery = gql`
   query orderForm {
     minicart @client {
-      orderForm {
-        cacheId
-        orderFormId
-        value
-        totalizers {
-          id
-          name
-          value
-        }
-        shippingData {
-          address {
-            id
-            neighborhood
-            complement
-            number
-            street
-            postalCode
-            city
-            reference
-            addressName
-            addressType
-          }
-          availableAddresses {
-            id
-            neighborhood
-            complement
-            number
-            street
-            postalCode
-            city
-            reference
-            addressName
-            addressType
-          }
-        }
-        clientProfileData {
-          email
-          firstName
-        }
-        storePreferencesData {
-          countryCode
-          currencyCode
-          timeZone
-        }
-      }
+      orderForm
     }
   }
 `
@@ -146,45 +20,7 @@ export const minicartOrderFormQuery = gql`
 export const minicartItemsQuery = gql`
   query {
     minicart @client {
-      items {
-        upToDate
-        id
-        name
-        imageUrl
-        detailUrl
-        skuName
-        quantity
-        sellingPrice
-        listPrice
-        seller
-        index
-        parentItemIndex
-        parentAssemblyBinding
-        options {
-          seller
-          quantity
-          assemblyId
-          id
-        }
-        assemblyOptions {
-          added {
-            item {
-              name
-              sellingPrice
-              quantity
-            }
-            normalizedQuantity
-            choiceType
-            extraQuantity
-          }
-          removed {
-            removedQuantity
-            initialQuantity
-            name
-          }
-          parentPrice
-        }
-      }
+      items
     }
   }
 `